SQL script for normalizing table 'UNF'
In this script the original database (with anomalies) is loaded in, then normalized by 1NF, 2NF and 3NF (further not needed in this example). 

In the Readme File you can see an ER diagram over the normalized database and also the script with description what I have done to normalize it. 

Input to UNF is the file denormalized-data.csv which you also can find in this git repository.